fix: fix pr comments - refctor
