+++ src/Common/Hooks/useNft/useNFTCollections.tsx
@@ -1,17 +1,7 @@
 import { useCallback } from "react"
-const allSettled = require("promise.allsettled")
 import { useThor } from "~Components"
 import { NonFungibleTokenCollection } from "~Model"
-import {
-    GithubCollectionResponse,
-    NftForContractResponse,
-    getCollectionInfo,
-    getContractAddresses,
-    getName,
-    getNftsForContract,
-    getSymbol,
-    getTokenURI,
-} from "~Networking"
+import { getCollectionInfo, getContractAddresses } from "~Networking"
 import {
     selectSelectedAccount,
     selectSelectedNetwork,
@@ -20,9 +10,9 @@ import {
     useAppDispatch,
     useAppSelector,
 } from "~Storage/Redux"
-import { fetchMetadata } from "./fetchMeta"
-import { NFTPlaceholder } from "~Assets"
 import { error } from "~Common/Logger"
+import { getNFTdataForContract, prepareCollectionData } from "./Helpers"
+import { useI18nContext } from "~i18n"
 
 /**
  * useNFTCollections
@@ -36,6 +26,7 @@ export const useNFTCollections = () => {
     const network = useAppSelector(selectSelectedNetwork)
     const selectedAccount = useAppSelector(selectSelectedAccount)
     const dispatch = useAppDispatch()
+    const { LL } = useI18nContext()
 
     const getCollections = useCallback(
         async (_page: number, _resultsPerPage: number = 10) => {
@@ -76,8 +67,8 @@ export const useNFTCollections = () => {
                     const { nftCollection } = await prepareCollectionData(
                         nft,
                         foundCollection,
-                        selectedAccount.address,
                         thor,
+                        LL.DATE_NOT_AVAILABLE(),
                     )
 
                     _nftCollections.push(nftCollection)
@@ -110,69 +101,8 @@ export const useNFTCollections = () => {
                 error("useNFTCollections", e)
             }
         },
-        [dispatch, network.type, selectedAccount.address, thor],
+        [LL, dispatch, network.type, selectedAccount.address, thor],
     )
 
     return { getCollections }
-}
-
-/*
- * Helper functions for useNFTCollections
- */
-const getNFTdataForContract = async (
-    contractsForNFTs: string[],
-    selectedAccountAddress: string,
-    _resultsPerPage: number,
-) => {
-    const NFTsForContractPromises: Promise<NftForContractResponse[]>[] = []
-
-    for (const contractAddress of contractsForNFTs) {
-        const nfts = getNftsForContract(
-            contractAddress,
-            selectedAccountAddress,
-            _resultsPerPage,
-        )
-        NFTsForContractPromises.push(nfts)
-    }
-
-    // resolve promises
-    const NFTsForContractRes = await allSettled(NFTsForContractPromises)
-
-    const nftData: NftForContractResponse[] = NFTsForContractRes.map(
-        (result: PromiseSettledResult<NftForContractResponse>) => {
-            if (result.status === "fulfilled") {
-                return result.value
-            }
-        },
-    )
-
-    return { nftData }
-}
-
-const prepareCollectionData = async (
-    nft: NftForContractResponse,
-    foundCollection: GithubCollectionResponse | undefined,
-    selectedAccountAddress: string,
-    thor: Connex.Thor,
-) => {
-    const _nft = nft.data[0]
-
-    const tokenURI = await getTokenURI(_nft.tokenId, _nft.contractAddress, thor)
-
-    const nftMeta = await fetchMetadata(tokenURI)
-
-    const nftCollection: NonFungibleTokenCollection = {
-        address: _nft.contractAddress,
-        name: await getName(_nft.contractAddress, thor),
-        symbol: await getSymbol(_nft.contractAddress, thor),
-        creator: foundCollection?.creator ?? "N/A",
-        description: foundCollection?.description ?? "",
-        icon: foundCollection?.icon
-            ? `https://vechain.github.io/nft-registry/${foundCollection?.icon}`
-            : nftMeta?.imageUrl ?? NFTPlaceholder,
-        balanceOf: nft.pagination.totalElements,
-        nfts: [],
-    }
-
-    return { nftCollection }
 }
